Reorganise code to scanner.go and conntrack.go
